+++ mapBuilder/www/js/components/LayerSelection/Slider.js
