+++ mapBuilder/www/js/main.js
@@ -26,7 +26,7 @@ import {always as alwaysCondition, shiftKeyOnly as shiftKeyOnlyCondition} from '
 import './modules/bottom-dock.js';
 
 import {LayerStore} from "./components/LayerStore";
-import {addElementToLayerArray} from "./modules/LayerSelection.js";
+import {addElementToLayerArray} from "./modules/LayerSelection/LayerSelection.js";
 
 // Extent on metropolitan France if not defined in mapBuilder.ini.php
 var originalCenter = [217806.92414447578, 5853470.637803803];
