+++ mapBuilder/www/js/modules/LayerSelection/LayerArray.js
