+++ mapBuilder/www/js/modules/LayerSelection/LayerElement.js
