+++ mapBuilder/www/js/modules/LayerTree/LayerTreeElement.js
