+++ mapBuilder/www/js/modules/LayerTree/LayerTreeFolder.js
