+++ mapBuilder/www/js/modules/LayerTree/LayerTreeLayer.js
